OpenFlowLib robustness
Removed several functions with NYIs that we don't need for the tutorial. I
believe some chaps want to fill these out eventually. That's fine. But,
instead of leaving NYIs everywhere, let's just have one big NYI in
Message.blit_message and Message.sizeof_body, as in this commit.That's fine. But,
instead of leaving NYIs everywhere, let's just have one big NYI in
Message.blit_message and Message.sizeof_body, as in this commit.That's fine. But,
instead of leaving NYIs everywhere, let's just have one big NYI in
Message.blit_message and Message.sizeof_body, as in this commit.That's fine. But,
instead of leaving NYIs everywhere, let's just have one big NYI in
Message.blit_message and Message.sizeof_body, as in this commit.