cexp: add int64 to value set, intro. cexp data type
Since the language now has non-policy values, there are opportunities
for type errors. In exchange the language now supports named constants,
though as of this commit, programs cannot use them effectively.
(Currently using an identifier to which a program has assigned a
constant will ensure that program produces a type error.)

The cexp data type is included to support additional expressions in the
future.