qt: add roundtrip test for OF0x01.PseudoPort
Somebody else should look at this commit. Because of they way that the
parse and marshal (i.e., make) functions work for PseudoPort, the code
manually tracks parameter of the Controller constructor in the
serialization datatype. This test therefore does not check whether or
not the artibrary instance generates valid data for that parameter.