qt: use PseudoPort.arbitrary_nc to ignore controller
Certain functions on in the OpenFlow0x01 API expect a pseudoPort but
do not handle the Controller variant properly. At some point the code
should discriminate on the different input domains at the type level,
but until then use this function to test.

With this commit the Enqueue case of the Action roundtrip test now
works, which was introduced but commented out in fa51a3f29.

Related to #5.