auto: initial commit NetKAT_Automoton module
This module will contain a translation to and from a representation of
policies as regular expressions over characters that are a link-free
policy sequenced with a link. From this, we'll derive an automaton
representing the policy.

As of now, the data types exist, as well as a translation from regex to
policy.