auto: working impl for (link-free policy; link)
The transformation definitely works for policies that are the sequence
of a link-free policy and a link. QuickCheck roundtrip tests are
included to verify, as well as a few specific regression tests that were
used to debug the previous version.

In addition, there are a few tests to test other "non-character"
policies, which suggest that the transformation is correct in general.
However, more testing is necessary.