auto: handle init choice with ingress policy
If the top-level constructor of a policy is Choice, then the previous
transformation would ignore the choice. This is because to_lf_policy
only checks for choice nodes when traversing to a new state. In that
case, there's a switch that can make the choice.

If there is no switch that can make the choice, then _all_ switches must
make the choice. So, return the choice as an ingress policy that the
compiler can implement as it chooses.