autoc: refactor creating choice policies
Define the foldl1_map policy, which is a combination of a left fold over
a non-empty list (i.e., it uses the first element as the accumulator)
and a map.

To create a choice policy, you map over the states creating modification
policies for each, and then combining them all with a Choice constructor
`foldl1_map pick mk_mod` does exactly that.