autoc: properly inject/eject packets from network
Provide some comments to clarify what's going. Briefly, this code
assumes all packets outside of the network are on vlan 0. The ingress
policy checks for this and then sets up the initial state of the
automaton. The egress policy checks for the final state (actually states
backwards reachable from the final state via epsilon transitions) and
ejects the packet from the network by setting the vlan back to 0.