autofix: Remove constructor from link_provider
A link_provider can accumulate link-free policies, but only via
sequential composition. That's because a link_provider represents a
policy that looks like this:

   (lfp ; r)

where lfp is a link-free policy, and r is a policy that's almost a
regular expression, when lfp is distributed across it. Associativity
matters here, so you can only extend lfp via sequential composition.

Related to #152, #155, and #158.