autofix: Better-working regex_of_policy
This commit changes the way that regex_of_policy works, and fixes some
bugs along the way. regex_of_policy now does two passes. First, it
draws boxes around link-free policies, and link-free policy, link pairs.
Then the algorithm works as it did before. There's one exception to
that, which is the Seq(NL, NL) case now works as it should.

The dump and run commands for automaton also now guards according to the
switch the policy should run on.