quickcheck: fix arbitrary_uint4 definition
The previous definition of arbitrary_unit4 was really arbitrary_uint3.
Correct the definition by introducing arbitrary_uint32_bits, which takes
the number of bits to randomly assign in the Int32.

The VLAN arbitrary instance does require three random bits, and for that
the code now uses arbitrary_uint32_bits directly.