reactive: Rework repesentation of stp_state (Closes #82)
ofp_port_state carries both a flag (whether or not the link is down) as
well as an enumeration (how the port is involved in a spanning tree
algorithm). The spanning tree states are mutually exclusive, so the code
now represents the possible states as an enumeration rather than a
collection of booleans.