Removed PingTest and fixed stupid _oasis conflict
PingTest was an early test that demonstated fast failover in action. It is
no longer needed. The real problem is that both PingTest and Test were
in the same directory but had different dependencies: PingTest required
Lwt and Test did not. But, since the _tags file adds dependencies by path,
Test would get linked to lwt even though it was not specified in the
_oasis file.

This is really terrible.

@seliopou you might enjoy the awful that is above