host: implement host discovery
Send ARP packets to the controller and infer from them where hosts are
attached to the network. This system assumes that the entire network is
under the management of the controller.

The implementation does not respond to any ARP requests. It simply
records information contained in them, and forwards the PACKET_IN event
to stage above.