intf: Rename ports_of_switch to ports_of_node
The current API seems incomplete, but exposes all the functions
necessary to work with hosts. Certain conventions need to be respected
by the user of the library, such as host destination port always being
set to 0 (this may be a bad assumption of a host is connected to
multiple switches), and then it seems to get you pretty far.