from_pipe: Determine the pipe from which a packet came
from_pipe takes a Local.t, which a controller may cache per switch, and
a packet, and determines the pipe that the packet belongs to. In general
this is not always possible. For a brief interval during switch setup,
packets may be sent to the controller not due to a policy. If a policy
makes use of different pipe name in non-disjoint parallel compositions,
then it may come from either, in which case the pipe the packet belongs
to is ambiguous.