from_pipe: Return all matching pipe
The user may have composed two applications that are interested in
packets that match identical filters, but are sent to different pipes.
from_pipe must account for this, and therefore return all pipes that
match the packet. It is the job of the runtime to then deliver the
packets to the proper pipe within the application, when possible.