from_pipe: Handle eval'd packets at physical locations
When a packet arrives at the controller, it's eval'd given the policy
currently installed on the switch. If that produces packets located at
physical locations, the controller will generate packet outs that will
perform the appropriate modifications in the form of actions.

Previously, the packet was not modified and no actions were inclued with
the packet_out.