vint: use int32 for portId and queueId
queue_id size is consistently 32 bits across OpenFlow versions. port_id,
however, is an int16 in OpenFlow 1.0 and an int32 in subsequent
versions. The compatibility layer will treat portIds as int32s and throw
a runtime error if a port number is used that is out of range for the
OpenFlow version the compatibility layer is using.

Note runtime checks should be occurring right now for reserved ranges of
port numbers anyways, so this does not introduce a point where errors
may occur.