vint: replace field/fieldVal with explicit actions
Rather than having a fieldVal type that must accommodate all the
different bit widths of fields, instead create an action for each field,
which explicitly determines the bit width of its associated value.

After this commit, it's no longer necessary to expose VInt in a public
API. It will remain in the code as a means to assist in casting between
various bit widths.