ports: Expose all common pseudoports in abstraction layer
Frenetic requires more than the previously exposed output options for
packets. In particular when compiling star, it may be necessary to send
a packet back out its input port specifically in the case of zero
iterations. While moving things around to InPort actions, it was
convenient to expose all the other shared pseudoports as well.

The point is to not hide too many things too early. An OpenFlow
abstraction layer should not hide core concepts of OpenFlow. It should
try and make them easier to use.