flowterp: Update to new ocaml-openflow API, fix tests
The failing tests were being caused by star test cases. There was no way
to represent in a flowtable the possibility of not going through the
star once, since ocaml-openflow did not expose the InPort pseudoport.
The new API does expose this, so a flowtable can be properly constructed
for star expressions.