app: sequence static katnetic static policy
It doesn't make much sense to run a static policy in parallel to
learning. Change the behavior of katnetic so that it sequences the
static policy before learning.

In the future, it may be useful to allow users to specify from the
command-line whether to sequence the static policy before or after
learning.