app: transfer updates in batches in create_async
The documentation for Pipe.transfer is out-of-date, and will only
transfer one item at a time. In the event that the size budget for
update queues is increased, it's necessary that this code do a batch
transfer in order to make full use of the size budget.

See janestreet/async_kernel#3 for more info.