net_updates: Txn module to handle RPC-like switch communication
This module exposes a send function that will return an Ivar.t that
will be filled with the response from the switch whe the controller
receives the response. The management of xids is hidden from the user.

Be sure to only use Txn.send with messages that will provoke the switch
to send a response. If you don't and you block on reading the returned
Ivar.t, then you'll be waiting until the switch disconnects.