update_fix: elminiate per-switch policy cache
The code for maintaining the specialized policy for each switch was
entangled in all the flowtable update code. Remove it for now so the
update code can be refactored in a sensible way. If there's a problem
with packet_in performance, then the cache can be reintroduced in a more
principled way.