batch: process policy updates in batches
Apply the last one in the Queue.t and disregard the rest. When used in
conjunction with consistent updates, this is a huge win as it eliminates
many interstitial policy states, reducing the number of versioned
flowtables that have to be installed. The result is that the network
converges much faster when starting up.