static: implement create_static more efficiently
The previous implementation of create_static was just a thin wrapper
around create, which kept the update and pkt_out pipes open even though
the app would never generate an update or a packet out. The new
implementation uses the lower-level raw_app API to construct the app and
in the process closes the update and pkt_out pipes.