http_add: add initialization stage
A user must specify a host, port, and path to create an http_app. The
path argument, together with the host and port, specifies the URI that
the application will send a POST request to in order to initialize a
application instance on the server. The application expects that the
Location header of the response will be populated with the path of the
application instance. This path will be used to send subsequent PUT
requests containing network events.