sw_id: Make Client_id Hashable.S rather than Unique_id
There is no need to expose the fact that the Client_id module satisfies
the siganture Unique_id, as users should never need to create additional
identifiers. In fact, the user may want to synthesize identifiers from
information gleaned from outside of the system. It's sufficient that the
module supports data structures such as hastables and sets.