multi: only report usable ports in switch features
The abstraction layer does not expose any of the data structures
necessary for the user to determine whether or not a port is live or
not. The abstraciton layer therefore needs to make this determination
for the user by filtering ports in the initial feature request, and
surfacing events when ports come up or down.