log: Update to use core/async.111.28.00 logging API
As of core/async.111.25.00, `Log.of_lazy` is no longer part of package's
public API. In 111.28.00, the `Log.level` call was allowing users of the
package to implement `of_lazy` without to manage the log level manually.
In addition, new optional arguments were introduced along with a
`Log.string` call. This is all now part of the public API of the
Async_OpenFlow module.

Closes #148.