flip: add module for simple runtime-configurable application
The Flip module supports building applications that can flip between two
different policies. Which policy is currently active is
runtime-configurable.

This module is written against the Raw_app interface which means that it
can be specialized both to a pred and policy application. It also means
that it should not be exported.