expose_raw: Reorg app interface into Policy module
The Async_NetKAT module has been reorganized into three submodules and
related functions. Creation functions are included in their respective
submodules, while combinators are in the top-level so they do not have
to be qualified for use. This should allow one to open the Async_NetKAT
module and get dropped into an environment for easy app creation and
composotion.