daft_punk: Use a Bus instead of a Pipe in the Flip app
When an app is used multiple times as a "sub-app", its handler is called
multiple times to create different callbacks. Each invocation of the
handler is passed a unique asynchronous update pipe. In order to support
reuse of the app, each asynchronous update caused by runtime
re-configuration must be written to each of these pipes. Previously, it
was only being written to one (whichever one the Async runtime chose to
perform first). Now, The ctl object maintains a bus that each occurrence
of the app subscribes to for asynchronous policy updates.