tdk: optimize sequential composition (part 2)
The branch case of the fold implements an if-then-else expression in the
Vlr. In the case that the true and false branches are equal, the
operation can just return the true branch and no further processing is
necessary.

This contributes a not insiginficant factor the performance improvement
of the compiler rewrite.