NetKAT QuickChecking now exercises variable order
I've modified the test suite to use the variable ordering heuristic when
QuickChecking the NetKAT axioms. Thus the tests not only test that the compiler
is correct, they also test that variable re-ordering is implemented correctly.

All tests do pass. In addition, I've verified that the compiler picks different
orders for each example, and in certain cases, when compiling two policies p
and q that should be equivalent, it picks different orders for for either side
of the equality. It still works, which is amazing.