New cache-clearing policy
In addition to completely preserving and clearing the cache, the
local compiler can now clear everything except an the data for
an existing FDD.

This fixes an issue with the controller, where FDD-operations are
used to compare the new and old FDD after compile is called.
Clearing the cache completely invalidates the old FDD. Using
the new `Preserve option, we can clear all the auxiliary data
used to create the old FDD but keep it intact.