Use the configure script to generate setup.data
To be consistent with other projects, there is now a configure script.
Just use that to generate setup.data.

I think the real fix is to have an _oasis-generated configure and Makefile that
just does the right thing.  This means we'd have to move the syntax extension
out of this repo.  I think the only reason we have a hand-written Makefile is
for the terrible sed hack.