Fixed the TCP hack. UDP now supported too.
Before the hack, the following policy:

 filter (ipSrc = 10.0.0.1 and ethTyp = 0x800); port := 200 |
 filter (not (ipSrc = 10.0.0.1 and ethTyp = 0x800)); port := 300

Would produce this table:

ipSrc = 10.0.0.1, ethTyp = 0x800 => port := 200
ipSrc = 10.0.0.1                 => port := 300
*                                => port := 300

Due to OpenFlow-weirdness, the second pattern is equivalent to just "*". The
hack was to patch it to this table:

ipSrc = 10.0.0.1, ethTyp = 0x800 => port := 200
ipSrc = 10.0.0.1, ethTyp = 0x800 => port := 300
*                                => port := 300

But, this hack is wrong! The second pattern is meant to be:

  ipSrc = 10.0.0.1, ethTyp != 0x800

However, the "!=" is inexpressible in OpenFlow.

Per OpenFlow, this predicate can never be true:

  ipSrc = 10.0.0.1 && ethTyp != 0x800

So, the right hack is to eliminate the second rule.

In this patch, the example policy will produce this table:

ipSrc = 10.0.0.1, ethTyp = 0x800 => port := 200
*                                => port := 300

@jcollard and @jnfoster read carefully!